feat(navbar): Swap out buttons depending on if user is logged in.
Resolve #66, Resolve #81
